+++ Controllers/MVBuddyListController.m
@@ -680,13 +680,13 @@ static MVBuddyListController *sharedInstance = nil;
 	[self _sortBuddiesAnimated:nil];
 }
 
-- (NSMenu *) tableView:(MVTableView *) tableView menuForTableColumn:(NSTableColumn *) tableColumn row:(int) row {
+- (NSMenu *) tableView:(MVTableView *) tableView menuForTableColumn:(NSTableColumn *) tableColumn row:(NSInteger) row {
 	if( tableView != buddies || row == -1 || row >= (int)[_buddyOrder count] ) return nil;
 	JVBuddy *buddy = [_buddyOrder objectAtIndex:row];
 	return [self _menuForBuddy:buddy];
 }
 
-- (NSString *) tableView:(MVTableView *) tableView toolTipForTableColumn:(NSTableColumn *) column row:(int) row {
+- (NSString *) tableView:(MVTableView *) tableView toolTipForTableColumn:(NSTableColumn *) column row:(NSInteger) row {
 	if( tableView != buddies || row == -1 || row >= (int)[_buddyOrder count] ) return nil;
 
 	NSMutableString *ret = [NSMutableString string];
@@ -766,7 +766,7 @@ static MVBuddyListController *sharedInstance = nil;
 #define curveFunction(t,p) ( pow( 1 - pow( ( 1 - t ), p ), ( p ? ( 1 / p ) : 0. ) ) )
 #define easeFunction(t) ( ( sin( ( t * M_PI ) - M_PI_2 ) + 1. ) / 2. )
 
-- (NSRect) tableView:(MVTableView *) tableView rectOfRow:(int) row defaultRect:(NSRect) defaultRect {
+- (NSRect) tableView:(MVTableView *) tableView rectOfRow:(NSInteger) row defaultRect:(NSRect) defaultRect {
 	if( _animating ) {
 		NSInteger oldPosition = [[_oldPositions objectAtIndex:row] intValue];
 		NSRect oldR = [tableView originalRectOfRow:oldPosition];
@@ -774,7 +774,7 @@ static MVBuddyListController *sharedInstance = nil;
 
 		float t = _animationPosition;
 
-		unsigned count = [_buddyOrder count];
+		NSUInteger count = [_buddyOrder count];
 		float rowPos = ( (float) row / (float) count );
 		float rowPosAdjusted = _viewingTop ? ( 1. - rowPos ) : rowPos;
 		float curve = 0.3;
