+++ Views/MVTableView.h
@@ -10,16 +10,16 @@
 - (BOOL) autosaveTableColumnHighlight;
 - (void) setAutosaveTableColumnHighlight:(BOOL) flag;
 
-- (NSRect) originalRectOfRow:(int) row;
+- (NSRect) originalRectOfRow:(NSInteger) row;
 @end
 
 @interface NSObject (MVTableViewDataSource)
-- (NSMenu *) tableView:(MVTableView *) view menuForTableColumn:(NSTableColumn *) column row:(int) row;
-- (NSString *) tableView:(MVTableView *) view toolTipForTableColumn:(NSTableColumn *) column row:(int) row;
+- (NSMenu *) tableView:(MVTableView *) view menuForTableColumn:(NSTableColumn *) column row:(NSInteger) row;
+- (NSString *) tableView:(MVTableView *) view toolTipForTableColumn:(NSTableColumn *) column row:(NSInteger) row;
 @end
 
 @interface NSObject (MVTableViewDelegate)
 - (void) clear:(id) sender;
-- (NSRect) tableView:(MVTableView *) tableView rectOfRow:(int) row defaultRect:(NSRect) defaultRect;
+- (NSRect) tableView:(MVTableView *) tableView rectOfRow:(NSInteger) row defaultRect:(NSRect) defaultRect;
 - (NSRange) tableView:(MVTableView *) tableView rowsInRect:(NSRect) rect defaultRange:(NSRange) defaultRange;
 @end
